ref(streams): Use f64 for watermark message_time
Address review on PR #314: align watermark message_time with the
codebase convention of f64 message timestamps so we no longer floor
sub-second precision when threading the value through native code,
serialization, and Python.

- Watermark, SerializableWatermark, PyWatermark, make_watermark_payload,
  Watermark::with_message_time and the Py<->Rust conversions now carry
  message_time as Option<f64> (Python float).
- WatermarkEmitter and Batch read PyAnyMessage/RawMessage timestamps
  directly as f64; the streaming_payload_timestamp_floor_secs and
  data_message_time_secs_from_routed helpers are removed.
- WatermarkCommitOffsets latency math becomes
  (current_epoch as f64) - oldest_message_time, clamped at 0 to absorb
  small clock skew.
- rust_streams.pyi and the build_watermark test helper expose
  message_time as float | None.

Drop the build_py_metrics_config docstring paragraph documenting the
new metric (per review).

+++ sentry_streams/src/sinks.rs
@@ -43,7 +43,7 @@ struct SerializableWatermark {
     committable: HashMap<String, Vec<HashMap<String, u64>>>,
     timestamp: u64,
     #[serde(default)]
-    message_time: Option<u64>,
+    message_time: Option<f64>,
 }
 
 impl From<Watermark> for SerializableWatermark {
